+++ src/main/java/ru/yandex/practicum/filmorate/model/Genre.java
@@ -5,11 +5,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/**
+ * Сущность которая описывает объект: Жанр фильма
+ */
+
 @Data
 @Builder(toBuilder = true)
 @NoArgsConstructor
 @AllArgsConstructor
 public class Genre {
-    private Integer id;
+    private Long id;
     private String name;
 }
