+++ src/main/java/ru/yandex/practicum/filmorate/model/Mpa.java
@@ -5,12 +5,16 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/**
+ * Сущность которая описывает объект: Возростной рейтинг фильма
+ */
+
 @Data
 @Builder(toBuilder = true)
 @NoArgsConstructor
 @AllArgsConstructor
 public class Mpa {
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
 }
