+++ src/main/java/ru/yandex/practicum/filmorate/storage/filmGenre/GenreStorage.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface GenreStorage {
 
-    Genre getGenreById(int id);
+    Genre getGenreById(long id);
 
     List<Genre> getAllGenre();
